requirements.txt
fitparse==1.2.0
garmin-fit-sdk==21.158.0
numpy==2.2.3
opencv-python==4.11.0.86
pillow==11.1.0
tqdm==4.67.1
